To represent the number of milliseconds since the epoch, need to use unsigned long long, rather than just unsigned long.

+++ examples/c++/readSBML.cpp
@@ -33,7 +33,7 @@ main (int argc, char* argv[])
   const char* filename   = argv[1];
   SBMLDocument* document;
   SBMLReader reader;
-  unsigned long start, stop;
+  unsigned long long start, stop;
 
   start    = getCurrentMillis();
   document = reader.readSBML(filename);
